Optimize the solution
Inspired by https://github.com/salvois/codejam/blob/master/src/codejam2017/finals/DiceStraight.java

The basic idea of the algorithm is to put all values on the faces of all dice in a list
sorted by value, storing the set of containing dice with each value. We then scan values
until consecutive values are present, and until we can  use a different die for each value.
When no free die is found for a value, we try to recursively shuffle the dice already used
for other values so that we can free a die for the new value.