Optimize the solution
Some ideas:

We want the fractional part of the percentage of each programming language to
be greater than or equal to 0.5, so that it rounds up during the final sum.
Thus, whenever we find a fractional part less than 0.5 we add the percentage
of one of the remaining people to try to make the fractional at least 0.5.
We give priority to percentages whose fractional part is as close as possible
to 0.5, in order to maximize the number of percentages we can bring above
the 0.5 threshold.
If we have no percentages whose fractional part is less than 0.5, we
add a new percentage for a new programming language, and repeat the process.
We don't touch percentages whose fractional part is already >= 0.5